Double -> java.math.BigDecimal

+++ src/main/kotlin/no/nav/pia/sykefravarsstatistikk/domene/Sykefraværsstatistikk.kt
@@ -1,10 +1,12 @@
 package no.nav.pia.sykefravarsstatistikk.domene
 
+import java.math.BigDecimal
+
 sealed interface Sykefraværsstatistikk {
     val årstall: Int
     val kvartal: Int
     val antallPersoner: Int
-    val tapteDagsverk: Double
-    val muligeDagsverk: Double
-    val prosent: Double
+    val tapteDagsverk: BigDecimal
+    val muligeDagsverk: BigDecimal
+    val prosent: BigDecimal
 }
